rag_excel/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
pandas>=2.0.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
faiss-cpu>=1.7.4
tiktoken>=0.5.1
tqdm>=4.66.1